Konflux Onboarding Prerequisites
Onboard the Jenkins WAR assembly to Konflux by packaging the core
jenkins.war into a container image. For now this is packaged in full
container image, complete with JDK development tools to facilitate
testing. In the future, the .war file should be packaged in either a
smaller "runtime" container image or distributed as an OCI artifact.